processing.stream_function returns a Field
It was previously returning an array. This change is to improve the
type soundness of the public API.

To avoid cyclic dependencies, a new `datatypes` is created to hold definitions
of several types that are ubiquitous in the code base.
