+++ docs/sources/apiref/datatypes.rst
@@ -1,0 +1,7 @@
+datatypes
+=========
+
+.. automodule:: stagpy.datatypes
+
+   .. autoclass:: Varf
+   .. autoclass:: Field
