+++ tests/test_phyvars.py
@@ -3,7 +3,7 @@ from stagpy import phyvars
 
 
 def test_dim():
-    allvars = chain(phyvars.FIELD.values(), phyvars.FIELD_EXTRA.values(),
+    allvars = chain(phyvars.FIELD.values(),
                     phyvars.RPROF.values(), phyvars.RPROF_EXTRA.values(),
                     phyvars.TIME.values(), phyvars.TIME_EXTRA.values())
     for var in allvars:
