+++ tests/test_processing.py
@@ -1,4 +1,4 @@
-from stagpy import processing
+from stagpy import processing, phyvars
 
 
 def test_dt_dt(sdat):
@@ -52,4 +52,5 @@ def test_energy_prof(step):
 
 def test_stream_function(step):
     psi = processing.stream_function(step)
-    assert psi.shape[1:3] == step.fields['v3'].values.shape[1:3]
+    assert psi.values.shape[1:3] == step.fields['v3'].values.shape[1:3]
+    assert psi.meta.dim in phyvars.SCALES
